# www/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fiat Lux api documentation",
        "description": "Fiat Lux School OpenApi documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/tests": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Returns a list of questions",
                "description": "Display a listing of the tests.",
                "operationId": "bd40e0627369d10b2c046ad5e00c747e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Test"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Unauthenticated",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "Update test question",
                "description": "Update the specified resource in storage.",
                "operationId": "88e60603f7480d209a190686b2ca4115",
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Unauthenticated",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Saves test question for test group",
                "description": "Store a newly created resource in storage.",
                "operationId": "4f39e46f9497955adc95a9d2502df728",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tests/group/{test_group_id}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Returns a list of questions by test group",
                "description": "Display a listing of the tests by test group id.",
                "operationId": "b4f42299c5c27f056a6541b04ebf3e82",
                "parameters": [
                    {
                        "name": "test_group_id",
                        "in": "path",
                        "description": "Test group id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Test"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Unauthenticated",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tests/group/{test_group_id}/quiz-questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Returns a list of questions by test group without is_correct flag for options",
                "description": "Display a listing of the tests by test group id for users to answer.",
                "operationId": "6a6c674b56e414bc9cf918e4371f4162",
                "parameters": [
                    {
                        "name": "test_group_id",
                        "in": "path",
                        "description": "Test group id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Test"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Unauthenticated",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreTestRequest": {
                "title": "Store Question request",
                "description": "Store Question request body data",
                "required": [
                    "question",
                    "test_group_id"
                ],
                "properties": {
                    "question": {
                        "title": "Test question",
                        "description": "Test question text",
                        "type": "string"
                    },
                    "test_group_id": {
                        "title": "Test group id",
                        "description": "Test group id, test is related",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Option": {
                "title": "Option Response model",
                "description": "Option Response model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Current test id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "text": {
                        "title": "Option text",
                        "description": "Option text",
                        "type": "string",
                        "format": "string"
                    },
                    "is_correct": {
                        "title": "Is correct option flag",
                        "description": "MArk if current option is correct for the question",
                        "type": "boolean",
                        "format": "boolean"
                    },
                    "test_id": {
                        "title": "Test id",
                        "description": "Test id for the current option",
                        "type": "integer",
                        "format": "int64"
                    },
                    "test": {
                        "$ref": "#/components/schemas/Test"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Option"
                }
            },
            "Test": {
                "title": "Test Response model",
                "description": "Test Response model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Current test id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "question": {
                        "title": "Test question",
                        "description": "Test question",
                        "type": "string",
                        "format": "string"
                    },
                    "test_group_id": {
                        "title": "Test group id",
                        "description": "Test group id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "test_group": {
                        "$ref": "#/components/schemas/TestGroup"
                    },
                    "options": {
                        "title": "List of options",
                        "description": "List of options for the current test",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Test"
                }
            },
            "TestGroup": {
                "title": "Test Group Response model",
                "description": "Test Group Response model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Current test group id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "group_name": {
                        "title": "Test group name",
                        "description": "Test group name",
                        "type": "string",
                        "format": "string"
                    },
                    "tests": {
                        "title": "Questions list",
                        "description": "List of test questions",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/Test"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TestGroup"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}